multi_agent_finance_assistant/requirements.txt
# Do NOT include faiss-cpu here if using conda
fastapi
uvicorn
streamlit
requests
beautifulsoup4
newspaper3k
yfinance
sentence-transformers
pyttsx3
openai-whisper
transformers
matplotlib
pydantic
